+++ app.js
@@ -1,41 +1,30 @@
-var express = require('express');
-var bodyParser = require('body-parser');
+var axios = require('axios');
 
-var app = express();
-var port = process.env.PORT || 3000;
+var bot = require('./bot.js');
 
-// body parser middleware
-app.use(bodyParser.urlencoded({ extended: true }));
+var { CLIENT_EVENTS, RTM_EVENTS, RtmClient } = require('@slack/client');
 
-// test route
-app.get('/', function (req, res) { res.status(200).send('Hello world!'); });
+var bot_token = 'xoxb-213951919538-fnBfDUHwsBdehp23Wbr0nZMI';
 
-app.post('/hello', function (req, res, next) {
-  var userName = req.body.user_name;
-  var botPayload = {
-    text : 'Hello ' + userName.toUpperCase() + ', welcome to TestMyBotHorizons Slack channel! I\'ll be your guide bitches!'
-  };
-  // Loop otherwise..
-  if (userName !== 'slackbot') {
-    return res.status(200).json(botPayload);
-  } else {
-    return res.status(200).end();
-  }
-});
+var rtm = new RtmClient(bot_token);
 
-app.post('/login', function (req, res, next) {
-  var userName = req.body.user_name;
-  var botPayload = {
-    text : 'Hello ' + userName.toUpperCase() + ', welcome to TestMyBotHorizons Slack channel! I\'ll be your guide bitches!'
-  };
-  // Loop otherwise..
-  if (userName !== 'slackbot') {
-    return res.status(200).json(botPayload);
-  } else {
-    return res.status(200).end();
-  }
-});
+rtm.on(CLIENT_EVENTS.RTM.AUTHENTICATED, (rtmStartData) => {
+  console.log(rtmStartData.self.name);
+})
 
-app.listen(port, function () {
-  console.log('Listening on port ' + port);
-});
+rtm.on(RTM_EVENTS.MESSAGE, (msg) => {
+  console.log(msg);
+  axios.get('https://api.api.ai/api/query?v=20150910&query=hi%20tiff&lang=en&sessionId=eabd5e33-630e-4ce3-894e-fc12ac3b0007&timezone=2017-07-17T14:01:03-0700',
+  {headers: {
+    'Authorization': 'Bearer 52da7b1243444adfb8d42bb5f51b07a3'
+  }})
+  .then((res) => {
+    console.log('result', res);
+    rtm.sendMessage('completed', msg.channel);
+  })
+  .catch((err) => {
+    console.log(err);
+  })
+})
+
+rtm.start();
