+++ bot.js
@@ -1,0 +1,43 @@
+// write express server in bot.js  and require it in here
+// sends link to user --> when you click on it it will send a request to express
+var express = require('express');
+var bodyParser = require('body-parser');
+
+var app = express();
+var port = process.env.PORT || 3000;
+
+// body parser middleware
+app.use(bodyParser.urlencoded({ extended: true }));
+
+// test route
+app.get('/', function (req, res) { res.status(200).send('Hello world!'); });
+
+app.post('/hello', function (req, res, next) {
+  var userName = req.body.user_name;
+  var botPayload = {
+    text : 'Hello ' + userName.toUpperCase() + ', welcome to TestMyBotHorizons Slack channel! I\'ll be your guide bitches!'
+  };
+  // Loop otherwise..
+  if (userName !== 'slackbot') {
+    return res.status(200).json(botPayload);
+  } else {
+    return res.status(200).end();
+  }
+});
+
+app.post('/login', function (req, res, next) {
+  var userName = req.body.user_name;
+  var botPayload = {
+    text : 'Hello ' + userName.toUpperCase() + ', welcome to TestMyBotHorizons Slack channel! I\'ll be your guide bitches!'
+  };
+  // Loop otherwise..
+  if (userName !== 'slackbot') {
+    return res.status(200).json(botPayload);
+  } else {
+    return res.status(200).end();
+  }
+});
+
+app.listen(port, function () {
+  console.log('Listening on port ' + port);
+});
